+++ docassemble_webapp/docassemble/webapp/update.py
@@ -455,7 +455,7 @@ def install_package(package):
         commands = ['pip', 'install']
         if disable_pip_cache:
             commands.append('--no-cache-dir')
-        commands.extend(['--quiet', '--prefix=' + PACKAGE_DIRECTORY, '--src=' + temp_dir, '--log-file=' + pip_log.name, '--force-reinstall', '--upgrade', saved_file.path + '.zip'])
+        commands.extend(['--quiet', '--prefix=' + PACKAGE_DIRECTORY, '--src=' + temp_dir, '--log-file=' + pip_log.name, '--upgrade', saved_file.path + '.zip'])
     elif package.type == 'git' and package.giturl is not None:
         if package.gitbranch is not None:
             branchpart = '@' + str(package.gitbranch)
@@ -473,12 +473,12 @@ def install_package(package):
             commands = ['pip', 'install']
             if disable_pip_cache:
                 commands.append('--no-cache-dir')
-            commands.extend(['--quiet', '--prefix=' + PACKAGE_DIRECTORY, '--src=' + temp_dir, '--force-reinstall', '--upgrade', '--log-file=' + pip_log.name, gitprefix + str(package.giturl) + gitsuffix + branchpart + '#egg=' + package.name + '&subdirectory=' + str(package.gitsubdir)])
+            commands.extend(['--quiet', '--prefix=' + PACKAGE_DIRECTORY, '--src=' + temp_dir, '--upgrade', '--log-file=' + pip_log.name, gitprefix + str(package.giturl) + gitsuffix + branchpart + '#egg=' + package.name + '&subdirectory=' + str(package.gitsubdir)])
         else:
             commands = ['pip', 'install']
             if disable_pip_cache:
                 commands.append('--no-cache-dir')
-            commands.extend(['--quiet', '--prefix=' + PACKAGE_DIRECTORY, '--src=' + temp_dir, '--force-reinstall', '--upgrade', '--log-file=' + pip_log.name, gitprefix + str(package.giturl) + gitsuffix + branchpart + '#egg=' + package.name])
+            commands.extend(['--quiet', '--prefix=' + PACKAGE_DIRECTORY, '--src=' + temp_dir, '--upgrade', '--log-file=' + pip_log.name, gitprefix + str(package.giturl) + gitsuffix + branchpart + '#egg=' + package.name])
     elif package.type == 'pip':
         if package.limitation is None:
             limit = ""
@@ -487,7 +487,7 @@ def install_package(package):
         commands = ['pip', 'install']
         if disable_pip_cache:
             commands.append('--no-cache-dir')
-        commands.extend(['--quiet', '--prefix=' + PACKAGE_DIRECTORY, '--src=' + temp_dir, '--force-reinstall', '--upgrade', '--log-file=' + pip_log.name, package.name + limit])
+        commands.extend(['--quiet', '--prefix=' + PACKAGE_DIRECTORY, '--src=' + temp_dir, '--upgrade', '--log-file=' + pip_log.name, package.name + limit])
     else:
         sys.stderr.write("Wrong package type\n")
         return 1, 'Unable to recognize package type: ' + package.name