+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [1.2.18] - 2020-12-22
+### Changed
+- Different startup mechanism for `celery`.
+- Removed `--force-reinstall` from `pip` invocation.
+
 ## [1.2.17] - 2020-12-17
 ### Added
 - The `chain` function from `itertools`.